# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: EduControl API
  version: '1.0'
  description: |
    Gestión de cursos, estudiantes y profesores.

    **Autorización.** Casi todo pide `Authorization: Bearer <token>`. El rol se
    lee de la base de datos en cada petición, no del claim del token: a quien
    degradan pierde el acceso de inmediato en vez de conservarlo hasta que
    caduque.

    **Propiedad.** El rol dice qué clase de usuario entra; de quién es el
    recurso lo decide el controlador leyéndolo. Un profesor solo toca sus
    cursos; un estudiante, sus matrículas.

    **Paginación.** Los listados devuelven `total`, `pagina`, `limite` y
    `paginas`, con un tope duro de 100 por página.
  license:
    name: MIT
    url: https://github.com/JohanParadaC/educontrol/blob/main/LICENSE
servers:
  - url: /
    description: El mismo origen que sirve la aplicación
...
security:
  - bearerAuth: []
paths:
...
  /api/cursos:
    post:
      operationId: crearCurso
      tags: [Cursos]
      summary: Crear un curso
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CursoEditable' }
      responses:
        '201':
          description: Curso creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  curso: { $ref: '#/components/schemas/Curso' }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/Prohibido' }
    get:
      operationId: listarCursos
      tags: [Cursos]
      summary: Catálogo de cursos
      description: >
        Los cursos archivados no salen para un estudiante; administración y
        profesorado sí los ven. `buscar` mira en nombre y descripción, y el
        texto se escapa antes de convertirlo en expresión regular.
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
        - name: profesor
          in: query
          description: '`me` para los que imparte quien pregunta, o un id'
          schema: { type: string }
        - name: buscar
          in: query
          schema: { type: string, maxLength: 100 }
      responses:
        '200':
          description: Página de cursos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginacion'
                  - type: object
                    properties:
                      cursos:
                        type: array
                        items: { $ref: '#/components/schemas/Curso' }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
  /api/cursos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: obtenerCurso
      tags: [Cursos]
      summary: Ficha de un curso
      description: >
        `matriculados` viene siempre. `estudiantes` **solo** si quien pregunta
        es el profesor de ese curso o un administrador: la clave se omite en
        vez de mandarse vacía, porque `[]` significaría "no hay ninguno". Un
        estudiante sabe cuántos son en su clase, no quiénes.
      responses:
        '200':
          description: El curso y su contexto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CursoDetalle' }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    put:
      operationId: actualizarCurso
      tags: [Cursos]
      summary: Editar un curso
      description: >
        Solo su profesor o un administrador. `cupoMaximo: null` quita el
        límite; el campo se borra en vez de guardarse a null.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CursoEditable' }
      responses:
        '200':
          description: Curso actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  curso: { $ref: '#/components/schemas/Curso' }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    delete:
      operationId: borrarCurso
      tags: [Cursos]
      summary: Borrar un curso
      description: Se lleva por delante sus inscripciones.
      responses:
        '200':
          description: Borrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  msg: { type: string }
                  inscripcionesEliminadas: { type: integer }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /api/cursos/{id}/estudiantes.csv:
    get:
      operationId: exportarEstudiantesCsv
      tags: [Cursos]
      summary: Exportar los matriculados a CSV
      description: >
        Solo el profesor del curso o un administrador. El fichero lleva BOM
        UTF-8 (para que Excel no rompa las tildes), saltos CRLF y entrecomillado
        RFC 4180. El nombre viaja en `Content-Disposition` dos veces: en ASCII
        y en RFC 5987, porque una cabecera HTTP no puede llevar tildes.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: El CSV
          headers:
            Content-Disposition:
              schema: { type: string }
              example: >-
                attachment; filename="Algebra_-_estudiantes.csv";
                filename*=UTF-8''%C3%81lgebra%20-%20estudiantes.csv
          content:
            text/csv:
              schema: { type: string }
              example: |
                Nombre,Correo,Matriculado el
                "Ruiz, Ana",ana@educontrol.com,2026-08-26
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /api/inscripciones:
    post:
      operationId: crearInscripcion
      tags: [Inscripciones]
      summary: Matricular a un estudiante
      description: >
        Por `estudianteId` (lo usa el panel de administración, que tiene la
        lista), por `correo` (lo usa el profesor, que no la tiene ni debe
        tenerla) o **sin destinatario**, y entonces se matricula quien lo pide.


        Quién puede matricular a quién: un administrador a cualquiera en
        cualquier curso; un profesor a cualquiera, pero solo en **los cursos
        que imparte**; un estudiante, solo a sí mismo. La comprobación va antes
        de resolver la cuenta destino, así que un 403 es idéntico exista ese
        correo o no.


        El duplicado lo decide el índice único de la base, no una consulta
        previa.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cursoId]
              properties:
                cursoId: { type: string, description: ObjectId del curso }
                estudianteId: { type: string, description: ObjectId del estudiante }
                correo: { type: string, format: email }
      responses:
        '201':
          description: Matrícula creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  inscripcion: { $ref: '#/components/schemas/Inscripcion' }
        '400':
          description: >
            Datos inválidos, la cuenta no es de un estudiante, o ya estaba
            matriculado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403':
          description: >
            No puedes matricular a esa persona en ese curso. Un estudiante solo
            se matricula a sí mismo; un profesor, solo en los cursos que
            imparte.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: El curso o la cuenta no existen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: >
            El curso está lleno, cerrado o archivado. No es un 400: el dato que
            manda el cliente está bien, lo que pasa es que el estado del recurso
            lo impide.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      estado: { type: string, enum: [cerrado, archivado] }
                      ocupadas: { type: integer }
                      cupoMaximo: { type: integer }
    get:
      operationId: listarInscripciones
      tags: [Inscripciones]
      summary: Listar matrículas
      description: >
        Devuelve lo que el rol permite ver: un estudiante las suyas, un profesor
        las de los cursos que imparte, un administrador todas. Los filtros se
        **cruzan** con esa regla: pedir algo fuera de lo permitido no amplía el
        resultado, lo vacía.
      parameters:
        - $ref: '#/components/parameters/Pagina'
        - $ref: '#/components/parameters/Limite'
        - name: curso
          in: query
          schema: { type: string }
        - name: estudiante
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Página de matrículas, con estudiante y curso poblados
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginacion'
                  - type: object
                    properties:
                      inscripciones:
                        type: array
                        items: { $ref: '#/components/schemas/Inscripcion' }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        `Authorization: Bearer <token>`. Es la única cabecera de sesión que se
        acepta: la antigua `x-token` se retiró, porque dos puertas para lo
        mismo son el doble de superficie.
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: ObjectId de 24 caracteres hexadecimales
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
    Pagina:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limite:
      name: limit
      in: query
      description: Tope duro de 100; pedir más se recorta en silencio
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
    NoAutenticado:
      description: Falta el token o no es válido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Prohibido:
      description: El rol o la propiedad del recurso no lo permiten
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NoEncontrado:
      description: No existe
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    DatosInvalidos:
      description: >
        Validación fallida. `errors` lleva un campo por cada problema, tal y
        como lo devuelve express-validator.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorValidacion' }
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        msg: { type: string }
    ErrorValidacion:
      type: object
      properties:
        ok: { type: boolean, example: false }
        errors:
          type: object
          additionalProperties:
            type: object
            properties:
              type: { type: string }
              msg: { type: string }
              path: { type: string }
              location: { type: string }
...
    Paginacion:
      type: object
      properties:
        ok: { type: boolean, example: true }
        total: { type: integer }
        pagina: { type: integer }
        limite: { type: integer }
        paginas: { type: integer }
...
    Usuario:
      type: object
      description: >
        El modelo renombra `_id` a `id` al serializar y nunca incluye la
        contraseña.
      properties:
        id: { type: string }
        nombre: { type: string }
        correo: { type: string, format: email }
        rol: { type: string, enum: [estudiante, profesor, admin] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Curso:
      type: object
      description: >
        El backend llama `nombre` a lo que la interfaz llama `titulo`; la
        traducción vive solo en el frontend.
      properties:
        _id: { type: string }
        nombre: { type: string, maxLength: 120 }
        descripcion: { type: string, maxLength: 500 }
        profesor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Usuario'
        cupoMaximo:
          type: integer
          minimum: 1
          description: Ausente es «sin límite» — nunca 0, que sería «no cabe nadie»
        estado: { type: string, enum: [abierto, cerrado, archivado], default: abierto }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CursoEditable:
      type: object
      properties:
        nombre: { type: string, maxLength: 120 }
        descripcion: { type: string, maxLength: 500 }
        profesor: { type: string, description: ObjectId de un usuario con rol profesor }
        cupoMaximo:
          type: integer
          minimum: 1
          nullable: true
          description: '`null` quita el límite; el campo se borra'
        estado: { type: string, enum: [abierto, cerrado, archivado] }
    CursoDetalle:
      type: object
      properties:
        ok: { type: boolean, example: true }
        curso: { $ref: '#/components/schemas/Curso' }
        matriculados: { type: integer }
        estudiantes:
          type: array
          description: >
            Solo para el profesor del curso o un administrador. Si no, la clave
            **no viene**: mandar `[]` diría «no hay ninguno», que es otra cosa.
          items: { $ref: '#/components/schemas/Usuario' }
    Inscripcion:
      type: object
      properties:
        _id: { type: string }
        estudiante:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Usuario'
        curso:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Curso'
        fecha:
          type: string
          format: date-time
          description: Se conserva por compatibilidad; para código nuevo, `createdAt`
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...